requirements.txt
streamlit
fastai==2.2
pillow==8.1
altair
pandas
